Updated AddEvent page
Made some changes on the add button and validation functionality. I decided to get rid of the disable attribute on the add button, since that interfered with the error message popping up when trying to submit empty fields. This way, the error message would be prioritized over the coloring of the button (as it is in the Sign Up page as well). The button will now always be green, except it will only submit if the text fields are valid and if not, it will throw an error message.